restructured organization
