+++ RGZcatalogCode/updateConsensus.py
@@ -6,7 +6,6 @@ import logging
 from pymongo import MongoClient
 import csv
 
-@profile
 def updateConsensus(csvPath):
 
     logging.basicConfig(filename='RGZcatalog.log', level=logging.DEBUG, format='%(asctime)s %(message)s')
